adapted overcooked_sim to work with the rest of the pipeline
